Refactor prediction step to eliminate multiprocessing
Previously, multiprocessing was utilized in the
prediction step, resulting in increased memory
consumption proportional
to the number of worker processes.
By refactoring the code to replace multiprocessing
with a simple for loop,
memory usage has been significantly reduced.
This change not only enhances efficiency
by eliminating the overhead of virtual memory
but also accelerates runtime performance.